+++ Socolin.RabbitMQ.Client/Options/Client/PerMessageTtlOption.cs
@@ -1,0 +1,12 @@
+namespace Socolin.RabbitMQ.Client.Options.Client
+{
+	public class PerMessageTtlOption
+	{
+		public readonly int? PerMessageTTl;
+
+		public PerMessageTtlOption(int? perMessageTTl)
+		{
+			PerMessageTTl = perMessageTTl;
+		}
+	}
+}