+++ Socolin.RabbitMQ.Client/Pipes/Client/Context/ClientPipeContextAction.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using RabbitMQ.Client;
-using Socolin.RabbitMQ.Client.Exceptions;
 
 namespace Socolin.RabbitMQ.Client.Pipes.Client.Context
 {
@@ -16,16 +15,6 @@ namespace Socolin.RabbitMQ.Client.Pipes.Client.Context
 		public ClientPipeContextAction(Func<IModel, ClientPipeContextAction, Task> action)
 		{
 			Action = action;
-		}
-
-		public T GetItemValue<T>(string key)
-		{
-			if (!Items.ContainsKey(key))
-				throw new MissingItemInContextException(key, typeof(T));
-			var obj = Items[key];
-			if (!(obj is T value))
-				throw new MissingItemInContextException(key, typeof(T));
-			return value;
 		}
 	}
 }