+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Allow specifying ExchangeName/RoutingKey when publishing a message as an alternative to QueueName.
 - Add new `CreateQueueOptionBuilder` to simplify how to create a queue with custom arguments.
 - Allow specifying `contextItems` when enqueueing a message. This databag is available through the `IPipeContext.Items` in pipes.
+- Add a new option to configure message expiration
 
 1.1